Extract dilbert module
